chore(tag): migrate to new api and update props
