+++ packages/admin-ui-docs/src/components/PageHeader/links.ts
@@ -6,6 +6,7 @@ export const PAGE_LINK: any = {
   FilterMultiple: { github: 'filter-multiple' },
   Filter: { github: 'filter' },
   Inline: { github: 'inline' },
+  Tag: { github: 'tag' },
   Alert: { github: 'components/Alert' },
   Anchor: { github: 'components/Anchor' },
   Avatar: { github: 'components/Avatar' },
@@ -45,7 +46,6 @@ export const PAGE_LINK: any = {
   Skeleton: { github: 'components/Skeleton' },
   Spinner: { github: 'components/Spinner' },
   Tabs: { github: 'components/Tabs' },
-  Tag: { github: 'components/Tag' },
   TextArea: { github: 'components/TextArea' },
   Text: { github: 'components/Text' },
   Toast: { github: 'components/Toast' },
