+++ packages/admin-ui/src/index.ts
@@ -18,6 +18,7 @@ export * from './text-input'
 export * from './box'
 export * from './search'
 export * from './number-input'
+export * from './tag'
 
 export * from './deprecated-set'
 
@@ -26,7 +27,6 @@ export * from './components/Label'
 export * from './components/Menu'
 export * from './components/Collapsible'
 export * from './components/CollapsibleGroup'
-export * from './components/Tag'
 export * from './components/Tooltip'
 export * from './components/Avatar'
 export * from './components/Anchor'
