+++ packages/admin-ui/src/inline/inline.stories.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import type { Meta } from '@storybook/react'
 
 import { Inline } from './inline'
-import { Tag } from '../components/Tag'
+import { Tag } from '../tag'
 import { Box } from '../box'
 
 export default {
