+++ packages/admin-ui/src/tag/index.tsx
@@ -1,0 +1,1 @@
+export * from './tag'
