+++ packages/admin-ui/src/tag/tag.test.tsx
@@ -1,10 +1,16 @@
 import React from 'react'
-import { render, axe } from '../../test-utils'
-import { IconPlus } from '@vtex/phosphor-icons'
-
-import { Tag } from './index'
+import { render, axe } from '../test-utils'
+import { Tag } from './tag'
 
 describe('Tag tests', () => {
+  it('should have overridable styles', () => {
+    const { getByTestId } = render(
+      <Tag label="tag-1" data-testid="tag" csx={{ bg: 'coral' }} />
+    )
+
+    expect(getByTestId('tag')).toHaveStyleRule('background', 'coral')
+  })
+
   it('should not have a11y violations', async () => {
     const { container } = render(
       <>
@@ -17,10 +23,6 @@ describe('Tag tests', () => {
         <Tag label="Rio de Janeiro" palette="teal" />
         <Tag label="Rio de Janeiro" palette="gray" />
         <Tag label="tag-1" />
-        <Tag label="tag-2" size="regular" />
-        <Tag label="tag-3" size="small" />
-        <Tag label="tag-4" handleDelete={() => {}} />
-        <Tag label="tag-5" icon={<IconPlus />} handleDelete={() => {}} />
       </>
     )
